folder structure improvements
